refactor(blorktools): handler package reorg
