+++ packages/blorktools/src/asset_debugger/util/data/upload/background-file-handler.js
@@ -1,7 +1,7 @@
-import { createClearButton } from '../../../../landing-page/landing-page';
-import { hidePreviewLoading, showPreviewLoading } from '../../../../loading-splash/preview-loading-splash';
-import { updateState } from '../../../state/scene-state';
-import { formatFileSize } from '../file-upload-manager';
+import { createClearButton } from '../../../landing-page/landing-page';
+import { hidePreviewLoading, showPreviewLoading } from '../../../loading-splash/preview-loading-splash';
+import { updateState } from '../../state/scene-state';
+import { formatFileSize } from './file-upload-manager';
 
 /**
  * Handle background image file upload
