+++ packages/blorktools/src/asset_debugger/util/data/upload/zip-handler.js
@@ -1,6 +1,6 @@
-import { formatFileSize } from "../file-upload-manager";
-import { TextureClassifier } from "../../texture-classifier";
-import { updateState } from "../../../state/scene-state";
+import { formatFileSize } from "./file-upload-manager";
+import { TextureClassifier } from "../texture-classifier";
+import { updateState } from "../../state/scene-state";
 
 const SUPPORTED_IMAGE_FORMATS = ['.png', '.jpg', '.jpeg', '.webp', '.tiff', '.tif'];
 const SUPPORTED_MODEL_FORMATS = ['.glb', '.gltf'];
